chore: move resources data mock to main DataMocks
